报表-员工推荐 报表-活动推荐

+++ yunya-report/yunya-middle-table/src/main/java/com/yunya/middletable/service/patient/BasePatientBiz.java
@@ -1,7 +1,9 @@
 package com.yunya.middletable.service.patient;
 
+import com.google.common.collect.Lists;
 import com.yunya.feign.report.domain.form.PullForm;
 import com.yunya.feign.report.domain.model.MessageModel;
+import com.yunya.feign.report.domain.vo.PatientTreatInfoVo;
 import com.yunya.framework.common.biz.BaseBiz;
 import com.yunya.framework.common.utils.StringHelper;
 import com.yunya.middletable.dao.patient.PatientBaseInfoMapper;
@@ -9,12 +11,14 @@ import com.yunya.middletable.dao.patient.PatientOriginMapper;
 import com.yunya.middletable.dao.patient.PatientPrepaymentsInfoMapper;
 import com.yunya.middletable.dao.report.BasePatientMapper;
 import com.yunya.middletable.dao.report.BasePatientMemberMapper;
+import com.yunya.middletable.dao.report.BaseTreatmentProcessMapper;
 import com.yunya.middletable.service.BaseTreatmentProcessBiz;
 import com.yunya.models.patient_central.PatientBaseInfo;
 import com.yunya.models.patient_central.PatientOrigin;
 import com.yunya.models.patient_central.PatientPrepaymentsInfo;
 import com.yunya.models.report.BasePatient;
 import com.yunya.models.report.BasePatientMember;
+import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,7 +27,10 @@ import org.springframework.transaction.annotation.Transactional;
 import tk.mybatis.mapper.entity.Example;
 
 import javax.annotation.Resource;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
@@ -42,15 +49,15 @@ import java.util.concurrent.Future;
 @Transactional(rollbackFor = Exception.class)
 public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
   /** 注入对象 */
-  @Autowired private PatientBaseInfoMapper patientBaseInfoMapper;
+  @Resource private PatientBaseInfoMapper patientBaseInfoMapper;
 
-  @Autowired private PatientOriginMapper patientOriginMapper;
+  @Resource private PatientOriginMapper patientOriginMapper;
 
-  @Autowired private BasePatientMemberBiz basePatientMemberBiz;
+  @Resource private BasePatientMemberBiz basePatientMemberBiz;
 
-  @Autowired private PatientPrepaymentsInfoMapper patientPrepaymentsInfoMapper;
+  @Resource private PatientPrepaymentsInfoMapper patientPrepaymentsInfoMapper;
 
-  @Autowired private BasePatientMemberMapper basePatientMemberMapper;
+  @Resource private BasePatientMemberMapper basePatientMemberMapper;
 
   /** 多线程 */
   @Resource(name = "customizeThreadPool")
@@ -72,7 +79,8 @@ public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
         addPrepaidInfo(patient);
         break;
       case 1:
-        mapper.updateByPrimaryKeySelective(patient);
+        assert patient != null;
+        mapper.updateByPrimaryKeySelective(getPatientInfo(patient));
         break;
       case 2:
         PatientBaseInfo patientBaseInfo = patientBaseInfoMapper.selectByPrimaryKey(patientId);
@@ -88,20 +96,63 @@ public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
   }
 
   /**
+   * 获取患者初末诊日期
+   * @param patient 患者信息
+   * @return 患者信息
+   */
+  private BasePatient getPatientInfo(BasePatient patient) {
+    if (patient != null){
+      // 初诊信息
+      PatientTreatInfoVo firstTreatInfo = mapper.selectFirstVisitInfo(patient.getPatientId());
+      if (null != firstTreatInfo) {
+        patient.setFirstVisitDate(getDateTime(firstTreatInfo.getTreatDate()));
+        patient.setFirstVisitDoctors(firstTreatInfo.getTreatDentistName());
+        patient.setFirstVisitOutpatient(firstTreatInfo.getTreatOutpatient());
+      }
+      // 末诊信息
+      PatientTreatInfoVo lastTreatInfo = mapper.selectLastVisitInfo(patient.getPatientId());
+      if (null != lastTreatInfo) {
+        patient.setLastVisitDate(getDateTime(lastTreatInfo.getTreatDate()));
+        patient.setLastVisitDoctors(lastTreatInfo.getTreatDentistName());
+        patient.setLastVisitOutpatient(lastTreatInfo.getTreatOutpatient());
+      }
+      return patient;
+    }
+   return null;
+  }
+
+  public Date getDateTime(String dateStr){
+    Date date = null;
+    //获得SimpleDateFormat类，我们转换为yyyy-MM-dd的时间格式
+    SimpleDateFormat sf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+    try {
+    //使用SimpleDateFormat的parse()方法生成Date
+     date = sf.parse(dateStr);
+    //打印Date
+      System.out.println(date);
+    } catch (ParseException e) {
+      e.printStackTrace();
+    }
+    return date;
+  }
+
+  /**
    * 添加患者信息是添加预付款信息
    *
    * @param patient 患者信息
    */
   public void addPrepaidInfo(BasePatient patient) {
-    PatientPrepaymentsInfo patientPrepaymentsInfo = new PatientPrepaymentsInfo();
-    patientPrepaymentsInfo.setPatientId(patient.getPatientId());
-    PatientPrepaymentsInfo patientPrepayments =
-        patientPrepaymentsInfoMapper.selectOne(patientPrepaymentsInfo);
-    if (patientPrepayments != null) {
-      BasePatientMember basePatientMember =
-          basePatientMemberBiz.getPatientMemberInfo(patientPrepayments.getId(), 1);
-      basePatientMemberMapper.deleteByPrimaryKey(basePatientMember);
-      basePatientMemberMapper.insert(basePatientMember);
+    if (patient != null){
+      PatientPrepaymentsInfo patientPrepaymentsInfo = new PatientPrepaymentsInfo();
+      patientPrepaymentsInfo.setPatientId(patient.getPatientId());
+      PatientPrepaymentsInfo patientPrepayments =
+              patientPrepaymentsInfoMapper.selectOne(patientPrepaymentsInfo);
+      if (patientPrepayments != null) {
+        BasePatientMember basePatientMember =
+                basePatientMemberBiz.getPatientMemberInfo(patientPrepayments.getId(), 1);
+        basePatientMemberMapper.deleteByPrimaryKey(basePatientMember);
+        basePatientMemberMapper.insert(basePatientMember);
+      }
     }
   }
 
@@ -215,4 +266,39 @@ public class BasePatientBiz extends BaseBiz<BasePatientMapper, BasePatient> {
   }
 
 
+  /**
+   * 批量更新患者信息
+   */
+  public void updPatientInfo() {
+    // 查询有过初诊的患者id
+   List<BasePatient> patientIdList  = mapper.selectPatientIdList();
+   if (patientIdList != null){
+     for (BasePatient  basePatient : patientIdList) {
+       // 初诊信息
+       PatientTreatInfoVo firstTreatInfo = mapper.selectFirstVisitInfo(basePatient.getPatientId());
+       if (null != firstTreatInfo) {
+         basePatient.setFirstVisitDate(getDateTime(firstTreatInfo.getTreatDate()));
+         basePatient.setFirstVisitDoctors(firstTreatInfo.getTreatDentistName());
+         basePatient.setFirstVisitOutpatient(firstTreatInfo.getTreatOutpatient());
+       }
+       // 末诊信息
+       PatientTreatInfoVo lastTreatInfo = mapper.selectLastVisitInfo(basePatient.getPatientId());
+       if (null != lastTreatInfo) {
+         basePatient.setLastVisitDate(getDateTime(lastTreatInfo.getTreatDate()));
+         basePatient.setLastVisitDoctors(lastTreatInfo.getTreatDentistName());
+         basePatient.setLastVisitOutpatient(lastTreatInfo.getTreatOutpatient());
+       }
+     }
+
+     List<List<BasePatient>> partitionLists = Lists.partition(patientIdList, 100);
+     CountDownLatch countDownLatch = new CountDownLatch(partitionLists.size());
+     long start = System.currentTimeMillis();
+     for (List<BasePatient> basePatientList: partitionLists) {
+       importExcelThreadPool.execute(() ->{
+
+       });
+     }
+   }
+
+  }
 }
